Move TxInfoStruct definition to business_logic/execution/objects
